Moved the Nero parser to its own package.
