--- lib/src/main/java/com/wjduquette/joe/nero/ErrorReporter.java
@@ -1,7 +1,0 @@
-package com.wjduquette.joe.nero;
-
-import com.wjduquette.joe.Trace;
-
-interface ErrorReporter {
-    void reportError(Trace trace);
-}
