+++ lib/src/main/java/com/wjduquette/joe/nero/parser/ErrorReporter.java
@@ -1,0 +1,7 @@
+package com.wjduquette.joe.nero.parser;
+
+import com.wjduquette.joe.Trace;
+
+public interface ErrorReporter {
+    void reportError(Trace trace);
+}
